Tex file parsing

+++ src/Parsers/MlbFile.h
@@ -18,7 +18,7 @@ namespace Sourcehold {
         class MlbFile : private Parser {
             /* Encoded in UTF-16LE */
             std::vector<std::wstring> field;
-            uint32_t num = 0;
+            uint16_t num = 0;
         public:
             MlbFile();
             ~MlbFile();
